Moved helper functions to lib/helpers.js
