+++ lib/index.js
@@ -1,20 +1,21 @@
 /* eslint-disable no-console */	
 
 'use strict';
-var _ = require('lodash');
-// var request = require('request');
-var md5 = require('md5');
 
 /**
  * Module dependencies.
  */
 
+var helper = require('./helpers');
+var request = require('request');
+var md5 = require('md5');
 var integration = require('@segment/analytics.js-integration');
+var apiBaseUrl = 'https://api.sailthru.com/';
 
 /**
  * Expose `Sailthru` integration.
  */
-var apiBaseUrl = 'https://api.sailthru.com/';
+
 
 var Sailthru = module.exports = integration('Sailthru')
   .global('Sailthru')
@@ -72,8 +73,8 @@ Sailthru.prototype.identify = function(identify) {
   var self = this;
   var props;
   var traits = identify.traits();
-  var id = getUserId(identify, props);
-  var email = getUserEmail(identify, props, traits);
+  var id = helper.getUserId(identify, props);
+  var email = helper.getUserEmail(identify, props, traits);
 
   // Config Options
   var integrations = identify.integrations();
@@ -96,14 +97,14 @@ Sailthru.prototype.identify = function(identify) {
     // onError: function() { console.log('Sign Up Error'); }
   };
 
-  payload.vars = filterCustomTraits(traits);
+  payload.vars = helper.filterCustomTraits(traits);
 
   if (defaultListName !== '') {
     var lists = payload.lists = {};
     lists[defaultListName] = 1;
   }
 
-  appendIdToPayload(payload, email, id);
+  helper.appendIdToPayload(payload, email, id);
 
   this._integration('userSignUp', payload);
 };
@@ -145,8 +146,8 @@ Sailthru.prototype.track = function(track) {
   var integrations = track.integrations();
   var props = track.properties();
   var traits = track.traits();
-  var id = getUserId(track, props);
-  var email = getUserEmail(track, props, traits);
+  var id = helper.getUserId(track, props);
+  var email = helper.getUserEmail(track, props, traits);
   var template = props && props.template
     || integrations && integrations.Sailthru && integrations.Sailthru.template;
   var payload = null;
@@ -162,7 +163,7 @@ Sailthru.prototype.track = function(track) {
       // onError: function() { console.log('Sign Up Confirmed Opt In Error'); }
     };
 
-    appendIdToPayload(payload, email, id);
+    helper.appendIdToPayload(payload, email, id);
 
     this._integration('userSignUpConfirmedOptIn', payload);
     break;
@@ -181,9 +182,9 @@ Sailthru.prototype.track = function(track) {
       // onError: function() { console.log('Custom Event Error'); }
     };
  
-    props = removePIIFromProperties(props);
-    payload.vars = flattenProperties(props);
-    appendIdToPayload(payload, email, id);
+    props = helper.removePIIFromProperties(props);
+    payload.vars = helper.flattenProperties(props);
+    helper.appendIdToPayload(payload, email, id);
 
     this._integration('customEvent', payload);
   }
@@ -215,8 +216,8 @@ Sailthru.prototype._addOrRemoveProduct = function(track, type) {
   var props = track.properties();
   var traits = track.traits();
   var context = track.context();
-  var email = getUserEmail(track, props, traits);
-  var id = getUserId(track, props);
+  var email = helper.getUserEmail(track, props, traits);
+  var id = helper.getUserId(track, props);
 
   var self = this;
   var options = self && self.options;
@@ -252,11 +253,10 @@ Sailthru.prototype._addOrRemoveProduct = function(track, type) {
   // .catch(function(error) {
   //   console.log(error);
   // });
-  // request
-  //   .get(url)
-  //   .on('error', function(err) { console.log('Error:'), console.error(err); })
-  //   .on('response', function(response) { console.log('Response:'), console.log(response); });
-  // var payload = addToCartPayload(product, productBaseUrl);
+  request
+    .get(url)
+    .on('error', function(err) { console.log('Error:'), console.error(err); })
+    .on('response', function(response) { console.log('Response:'), console.log(response); });
 
   // For testing
   var cart = {
@@ -270,9 +270,9 @@ Sailthru.prototype._addOrRemoveProduct = function(track, type) {
   // For testing
 
   var product = [props];
-  var item = mapItems(product, productBaseUrl)[0];
-  var payload = updateCart(item, cart, type);
-  appendIdToPayload(payload, email, id);
+  var item = helper.mapItems(product, productBaseUrl)[0];
+  var payload = helper.updateCart(item, cart, type);
+  helper.appendIdToPayload(payload, email, id);
 
   if (reminderTemplate !== '' && reminderTime !== '') { 
     payload.reminder_template = reminderTemplate;
@@ -281,7 +281,6 @@ Sailthru.prototype._addOrRemoveProduct = function(track, type) {
 
   this._integration('addToCart', payload);
 };
-
 
 /**
  * orderUpdated events map to addToCart Sailthru event
@@ -294,8 +293,8 @@ Sailthru.prototype.orderUpdated = function(track) {
   var props = track.properties();
   var traits  = track.traits();
   var context = track.context();
-  var id = getUserId(track, props);
-  var email = getUserEmail(track, props, traits);
+  var id = helper.getUserId(track, props);
+  var email = helper.getUserEmail(track, props, traits);
   var pageUrl = context && context.page && context.page.url;
   var products = props && props.products;
 
@@ -311,20 +310,20 @@ Sailthru.prototype.orderUpdated = function(track) {
     || options && options.reminderTime;
   var productBaseUrl = options && options.productBaseUrl;
 
-  var items = mapItems(products, pageUrl, productBaseUrl);
+  var items = helper.mapItems(products, pageUrl, productBaseUrl);
 
   var payload = {
     items: items,
-    adjustments: getAdjustments(track, props),
+    adjustments: helper.getAdjustments(track, props),
     incomplete: 1
     // onSuccess: function() { console.log('addToCart Successful'); },
     // onError: function() { console.log('addToCart Error'); }
   };
 
   delete props.products;
-  props = removePIIFromProperties(props);
-  payload.props = flattenProperties(props);
-  appendIdToPayload(payload, email, id);
+  props = helper.removePIIFromProperties(props);
+  payload.vars = helper.flattenProperties(props);
+  helper.appendIdToPayload(payload, email, id);
 
   if (reminderTemplate !== '' && reminderTime !== '') { 
     payload.reminder_template = reminderTemplate;
@@ -345,8 +344,8 @@ Sailthru.prototype.orderCompleted = function(track) {
   var props = track.properties();
   var traits  = track.traits();
   var context = track.context();
-  var id = getUserId(track, props);
-  var email = getUserEmail(track, props, traits);
+  var id = helper.getUserId(track, props);
+  var email = helper.getUserEmail(track, props, traits);
   var products = props && props.products;
   var pageUrl = context && context.page && context.page.url;
 
@@ -359,19 +358,19 @@ Sailthru.prototype.orderCompleted = function(track) {
     || options && options.sendTemplate;
   var productBaseUrl = options && options.productBaseUrl;
 
-  var items = mapItems(products, pageUrl, productBaseUrl);
+  var items = helper.mapItems(products, pageUrl, productBaseUrl);
 
   var payload = {
     items: items,
-    adjustments: getAdjustments(track, props)
+    adjustments: helper.getAdjustments(track, props)
     // onSuccess: function() { console.log('Purchase Successful'); },
     // onError: function() { console.log('Purchase Error'); }
   };
 
   delete props.products;
-  props = removePIIFromProperties(props);
-  payload.vars = flattenProperties(props);
-  appendIdToPayload(payload, email, id);
+  props = helper.removePIIFromProperties(props);
+  payload.vars = helper.flattenProperties(props);
+  helper.appendIdToPayload(payload, email, id);
 
   if (sendTemplate !== '') { payload.send_template = sendTemplate; }
 
@@ -402,238 +401,3 @@ Sailthru.prototype._integration = function(event, payload) {
   console.log(payload);
   window.Sailthru.integration(event, payload);
 };
-
-// Helper Functions:
-
-/**
- * Takes an array of products and maps them to an array of Items in Sailthru format
- *
- * @param {Array} products - array of products from event.properties.products
- * @param {String} productBaseUrl - default product url to be used if a product does not have a given url
- */
-
-function mapItems(products, pageUrl, productBaseUrl) {
-  var items = [];
-  if (products) {
-    products.forEach(function(product) {
-      var productId = product && product.product_id || product && product.id || product && product.sku || null;
-      product.price = formatPrice(product.price);
-      var item = {
-        qty: product && product.quantity || 1,
-        title: product && product.name || '',
-        price: product && product.price || 0,
-        id: productId,
-        url: product && product.url || pageUrl + '/' + productId || productBaseUrl + '/' + productId,
-        images: {
-          full: {
-            url: product && product.image_url || ''
-          },
-          thumb: {
-            url: product && product.image_url_thumb || ''
-          }
-        }
-      };
-
-      item.vars = filterProductCustomProperties(product); 
-      items.push(item);
-    });
-  }
-  return items;
-}
-
-function updateCart(currentItem, cart, type) {
-  if (cart.items.some(function(item) { return item.id === currentItem.id; })) {
-    console.log('Item found in cart');
-    var itemInCart = cart.items.find(function(item) { return item.id === currentItem.id; });
-    console.log('itemInCart:');
-    console.log(itemInCart);
-    var newQty;
-    if (type === 'remove') {
-      newQty = itemInCart.qty - currentItem.qty;
-    } else {
-      newQty = itemInCart.qty + currentItem.qty;
-    }
-    console.log('newQty:');
-    console.log(newQty);
-    itemInCart.vars = Object.assign({}, itemInCart.vars, currentItem.vars);
-    itemInCart.qty = newQty;
-    itemInCart.quantity = newQty;
-    console.log('itemInCart after update:');
-    console.log(itemInCart);
-    cart.items = cart.items.filter(function(item) { return item.qty > 0; });
-    return cart;
-  }
-  console.log('Item not found in cart');
-  if (type !== 'remove') {
-    cart.items.push(currentItem);
-  }
-  return cart;
-}
-
-/**
- * Retrieves the value to be assigned as the user's ID. userId will be used if available, otherwise anonymousId will be used.
- *
- * @param {Object} event - event which contains anonymousId and possibly userId
- * @param {Object} props - properties object which may contain userId or anonymousId
- */
-
-function getUserId(event, props) {
-  return event.userId() || props && props.userId || event.anonymousId() || props && props.anonymousId;
-}
-
-/**
- * Retrieves the value to be assigned as the user's email. Searches for email value at the event level first, then inside traits and properties in that order.
- *
- * @param {Object} event - event object
- * @param {Object} props - event properties
- * @param {Object} traits - event traits
- */
-
-function getUserEmail(event, props, traits) {
-  return event.email() || traits && traits.email || props && props.email;
-}
-
-/**
- * Appends email to event payload if available. If email is undefined, either userId or anonymousId will be assigned as extid.
- *
- * @param {Object} payload - event payload to which user data wil be appended
- * @param {String} email - user's email taken from event, event.traits, or event.properties
- * @param {String} id - user's id taken from event.userId or event.anonymousId
- */
-
-function appendIdToPayload(payload, email, id) {
-  if (email) { 
-    payload.email = email;
-    return; 
-  }
-  payload.id = id,
-  payload.key = 'extid';
-}
-
-/**
- * Removes common user traits
- *
- * @param {Object} traits - user traits
- */
-
-function filterCustomTraits(traits) {
-  return _.omit(traits, [
-    'email',
-    'defaultListName',
-    'optout_email',
-    'id',
-    'userId',
-    'annonymousId'
-  ]);
-}
-
-/**
- * Removes common product properties
- *
- * @param {Object} properties - product properties
- */
-
-function filterProductCustomProperties(properties) {
-  return _.omit(properties, [
-    'url',
-    'value',
-    'quantity',
-    'name',
-    'price',
-    'product_id',
-    'id',
-    'image_url',
-    'image_url_thumb'
-  ]);
-}
-
-/**
- * Removes Personally Identifiable Information from properties
- *
- * @param {Object} properties - event
- */
-
-function removePIIFromProperties(properties) {
-  return _.omit(properties, [
-    'email',
-    'firstName',
-    'lastName',
-    'gender',
-    'city',
-    'country',
-    'phone',
-    'state',
-    'zip',
-    'birthday'
-  ]);
-}
-
-/**
- * Removes Personally Identifiable Information from properties
- * Adapted from: https://stackoverflow.com/a/53739792
- * 
- * @param {Object} props - event properties
- */
-
-function flattenProperties(props) {
-  var toReturn = {};
-
-  for (var i in props) {
-    if (!props.hasOwnProperty(i)) continue;
-
-    if (typeof props[i] === 'object' && props[i] !== null) {
-      var flatObject = flattenProperties(props[i]);
-      for (var x in flatObject) {
-        if (!flatObject.hasOwnProperty(x)) continue;
-
-        toReturn[i + '_' + x] = flatObject[x];
-      }
-    } else {
-      toReturn[i] = props[i];
-    }
-  }
-  return toReturn;
-}
-
-/**
- * Format tax, shipping, and discount properties as price adjustments
- * 
- * @param {Object} event - event object
- * @param {Object} props - event properties
- */
-
-function getAdjustments(event, props) {
-  var tax = event.tax() || props && props.tax;
-  var shipping = event.shipping() || props && props.shipping;
-  var discount = event.discount() || props && props.discount;
-  var adjustments = [];
-
-  if (tax) {
-    adjustments.push({
-      title: 'tax',
-      price: formatPrice(tax)
-    });
-  }
-  if (shipping) {
-    adjustments.push({
-      title: 'shipping',
-      price: formatPrice(shipping)
-    });
-  }
-  if (discount) {
-    adjustments.push({
-      title: 'discount',
-      price: -formatPrice(discount) 
-    });
-  }
-
-  return adjustments;
-}
-
-/**
- * Convert price in dollars to price in cents
- * 
- * @param {Object} price - price in dollars
- */
-
-function formatPrice(price) { return Math.round(price * 100); } // absolute value
